third_party/blink/web_tests/fast/forms/suggestion-picker/date-suggestion-picker-min-max-attribute.html

<!DOCTYPE html>
<meta name=fuzzy content="maxDifference=0-3; totalPixels=0-1000">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../resources/picker-common.js"></script>
<script src="resources/suggestion-picker-common.js"></script>

<body style="background-color: #bbbbbb;">
<input type=date value="2012-12-24" min="2012-01-02" max="2012-01-08" list=suggestions>
<datalist id=suggestions>
  <option>2012-01-01</option>
  <option>2012-01-02</option>
  <option>2012-01-03</option>
  <option>2012-01-04</option>
  <option>2012-01-05</option>
  <option>2012-01-06</option>
  <option>2012-01-07</option>
  <option>2012-01-08</option>
  <option>2012-01-09</option>
  <option>2012-01-10</option>
</datalist>

<script>
window.onload = () => {
  const element = document.querySelector('input');
  promise_test(() => {
    return openPicker(element)
    .then(() => {
      internals.pagePopupWindow.focus();

      assert_equals(entryValues().length, 8);
      assert_equals(entryValues()[0], '2012-01-02');
      assert_equals(entryValues()[1], '2012-01-03');
      assert_equals(entryValues()[2], '2012-01-04');
      assert_equals(entryValues()[3], '2012-01-05');
      assert_equals(entryValues()[4], '2012-01-06');
      assert_equals(entryValues()[5], '2012-01-07');
      assert_equals(entryValues()[6], '2012-01-08');
      assert_equals(entryValues()[7], '@openCalendarPicker');

      return attemptToClosePicker(element);
    });
  },'min/max attributes filter out suggestions');
};
</script>
